Recommender/requirements.txt
requests~=2.32.3
rdflib~=7.0.0
neo4j~=5.20.0
Flask~=3.0.3
python-dotenv~=1.0.1